requirements.txt
python-telegram-bot>=20.0
gspread>=5.10.0
google-auth>=2.20.0
python-dotenv>=1.0.0
pytz>=2023.3
